Code_on_track-daksh(ai)/backend/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
python-dateutil
flask-cors
scikit-learn
pandas
scipy
APScheduler
ortools
opencv-python-headless
